+++ includes/slm_plugin_core.php
@@ -98,6 +98,7 @@ if (null !== SLM_Helper_Class::slm_get_option('slm_woo') && SLM_Helper_Class::sl
     require_once( SLM_ADMIN  . 'includes/woocommerce/wc-software-license-manager.php');
     // support for meta boxes
     require_once( SLM_LIB . 'slm-meta-boxes.php');
+    require_once( SLM_LIB . 'slm-wc-order-post-type.php');
 }
 
 if (null !== SLM_Helper_Class::slm_get_option('slm_wpestores') && SLM_Helper_Class::slm_get_option('slm_wpestores') == 1) {